fixed visibility of add item delete item and delete list to non-owner. Also cheanged delete button for non-owner to mark purchased button which changes purchased status. Also made purchased status not viewable to owner

+++ project/app/controllers/list_items_controller.rb
@@ -106,8 +106,17 @@ class ListItemsController < ApplicationController
 
   def destroy
     item = ListItem.find(params[:id])
-    item.destroy
-    flash[:warning] = 'Item Deleted'
+    if item.list.user_id != current_user.id
+      if item.purchased.zero?
+        item.purchased = 1
+      else
+        item.purchased = 0
+      end
+      item.save
+    else
+      item.destroy
+      flash[:warning] = 'Item Deleted'
+    end
     redirect_to request.referrer
   end
 end
